move proc timeout functions and tests to analyzer
Process timeout handling is only done in the analyzer part.
Related functions and tests are moved to the analyzer package.

tu_collector import was moved so it is not required to run the
unit tests.
